Moved item layer entities to their own subpackage
Refactored calls to the new package
