+++ SolStandard/Entity/General/Item/Key.cs
@@ -8,26 +8,39 @@ using SolStandard.Map.Elements;
 using SolStandard.Utility;
 using SolStandard.Utility.Assets;
 
-namespace SolStandard.Entity.General
+namespace SolStandard.Entity.General.Item
 {
-    public class Currency : TerrainEntity, IActionTile
+    public class Key : TerrainEntity, IItem, IActionTile
     {
-        public const string CurrencyAbbreviation = "G";
-
-        public int Value { get; private set; }
+        public string UsedWith { get; private set; }
         public int[] Range { get; private set; }
 
-        public Currency(string name, string type, IRenderable sprite, Vector2 mapCoordinates,
-            Dictionary<string, string> tiledProperties, int value, int[] range) :
+        public Key(string name, string type, IRenderable sprite, Vector2 mapCoordinates,
+            Dictionary<string, string> tiledProperties, string usedWith, int[] range) :
             base(name, type, sprite, mapCoordinates, tiledProperties)
         {
-            Value = value;
+            UsedWith = usedWith;
             Range = range;
+        }
+
+        public IRenderable Icon
+        {
+            get { return Sprite; }
         }
 
         public UnitAction TileAction()
         {
-            return new PickUpCurrencyAction(this);
+            return new PickUpItemAction(this, MapCoordinates);
+        }
+
+        public UnitAction UseAction()
+        {
+            return new ToggleLockAction(this);
+        }
+
+        public UnitAction DropAction()
+        {
+            return new DropItemAction(this);
         }
 
         public override IRenderable TerrainInfo
@@ -47,11 +60,11 @@ namespace SolStandard.Entity.General
                                 (CanMove) ? PositiveColor : NegativeColor)
                         },
                         {
-                            new RenderText(AssetManager.WindowFont, "Value: " + Value + CurrencyAbbreviation),
+                            new RenderText(AssetManager.WindowFont, "Used with: " + UsedWith),
                             new RenderBlank()
                         }
                     },
-                    1
+                    3
                 );
             }
         }
