+++ SolStandard/Entity/Unit/Actions/Terrain/ToggleLockAction.cs
@@ -2,6 +2,7 @@
 using SolStandard.Containers;
 using SolStandard.Containers.Contexts;
 using SolStandard.Entity.General;
+using SolStandard.Entity.General.Item;
 using SolStandard.Map.Elements;
 using SolStandard.Map.Elements.Cursor;
 using SolStandard.Utility.Assets;
