+++ SolStandard/Utility/Events/IncreaseUnitGoldEvent.cs
@@ -1,5 +1,5 @@
 ﻿using SolStandard.Containers.Contexts;
-using SolStandard.Entity.General;
+using SolStandard.Entity.General.Item;
 using SolStandard.Utility.Assets;
 
 namespace SolStandard.Utility.Events
