+++ SolStandard/Utility/Events/PickUpCurrencyEvent.cs
@@ -1,6 +1,6 @@
 ﻿using SolStandard.Containers;
 using SolStandard.Containers.Contexts;
-using SolStandard.Entity.General;
+using SolStandard.Entity.General.Item;
 using SolStandard.Map;
 using SolStandard.Utility.Assets;
