+++ SolStandard/Utility/Events/TakeSpoilsEvent.cs
@@ -1,7 +1,7 @@
 ﻿using SolStandard.Containers;
 using SolStandard.Containers.Contexts;
 using SolStandard.Entity;
-using SolStandard.Entity.General;
+using SolStandard.Entity.General.Item;
 using SolStandard.Map;
 using SolStandard.Utility.Assets;
